+++ src/app/api/articles/uploadImage/route.ts
@@ -1,5 +1,10 @@
 import { createResponse } from "@/lib/session";
-import { generateToken, isAuth } from "@/lib/utils";
+import {
+  generateToken,
+  isAuth,
+  signInFirebase,
+  signOutFirebase,
+} from "@/lib/utils";
 import { NextRequest } from "next/server";
 import {
   getDownloadURL,
@@ -7,7 +12,13 @@ import {
   uploadBytes,
   deleteObject,
 } from "firebase/storage";
-import { storage } from "@/lib/firebase";
+import { app, storage } from "@/lib/firebase";
+import {
+  Auth,
+  getAuth,
+  signInWithEmailAndPassword,
+  signOut,
+} from "firebase/auth";
 
 export async function POST(request: NextRequest) {
   // Create response
@@ -15,27 +26,42 @@ export async function POST(request: NextRequest) {
   // Create session
   const { currentUser } = await isAuth(request, response);
   if (currentUser?.role === "ADMIN" && currentUser.verified) {
-    const formData = await request.formData();
-    // @ts-ignore
-    const image: File = formData.get("image")!;
-    const filename = generateToken() + ".jpg";
     // Upload to firebase storage
-    const storageRef = ref(storage, `images/${filename}`);
-    const { metadata, ref: ubRef } = await uploadBytes(storageRef, image);
-    const url = await getDownloadURL(storageRef);
-
-    return createResponse(
-      response,
-      JSON.stringify({
-        success: 1,
-        file: {
-          url: url,
-        },
-      }),
-      {
-        status: 200,
-      }
-    );
+    const auth = getAuth(app);
+    const userId = await signInFirebase(auth);
+    if (userId) {
+      const formData = await request.formData();
+      // @ts-ignore
+      const image: File = formData.get("image")!;
+      const filename = generateToken() + ".jpg";
+      const storageRef = ref(storage, `images/${filename}`);
+      const { metadata, ref: ubRef } = await uploadBytes(storageRef, image);
+      const url = await getDownloadURL(storageRef);
+      await signOutFirebase(auth);
+      console.log("URL", url);
+      return createResponse(
+        response,
+        JSON.stringify({
+          success: 1,
+          file: {
+            url: url,
+          },
+        }),
+        {
+          status: 200,
+        }
+      );
+    } else {
+      return createResponse(
+        response,
+        JSON.stringify({
+          success: 0,
+        }),
+        {
+          status: 500,
+        }
+      );
+    }
   } else {
     return createResponse(
       response,
@@ -55,22 +81,40 @@ export async function DELETE(request: NextRequest) {
   // Create session
   const { currentUser } = await isAuth(request, response);
   if (currentUser?.role === "ADMIN" && currentUser.verified) {
-    const { filename } = await request.json();
-    // Delete image from firebase
-    const storageRef = ref(storage, `images/${filename}`);
-    await deleteObject(storageRef);
-    return createResponse(
-      response,
-      JSON.stringify(
-        JSON.stringify({
-          success: true,
-          message: "Deleted image successfully.",
-        })
-      ),
-      {
-        status: 200,
-      }
-    );
+    const auth = getAuth(app);
+    const userId = await signInFirebase(auth);
+    if (userId) {
+      const { filename } = await request.json();
+      // Delete image from firebase
+      const storageRef = ref(storage, `images/${filename}`);
+      await deleteObject(storageRef);
+      await signOutFirebase(auth);
+      return createResponse(
+        response,
+        JSON.stringify(
+          JSON.stringify({
+            success: true,
+            message: "Deleted image successfully.",
+          })
+        ),
+        {
+          status: 200,
+        }
+      );
+    } else {
+      return createResponse(
+        response,
+        JSON.stringify(
+          JSON.stringify({
+            success: false,
+            message: "SystemError: Failed to delete.",
+          })
+        ),
+        {
+          status: 500,
+        }
+      );
+    }
   } else {
     return createResponse(
       response,
